WIP on moving logging events to custom serializer
