+++ subprojects/logging/src/main/java/org/gradle/internal/logging/serializer/SpanSerializer.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.logging.serializer;
+
+import org.gradle.internal.logging.events.StyledTextOutputEvent;
+import org.gradle.internal.logging.text.StyledTextOutput;
+import org.gradle.internal.serialize.Decoder;
+import org.gradle.internal.serialize.Encoder;
+import org.gradle.internal.serialize.Serializer;
+
+public class SpanSerializer implements Serializer<StyledTextOutputEvent.Span> {
+    private final Serializer<StyledTextOutput.Style> styleSerializer;
+
+    public SpanSerializer(Serializer<StyledTextOutput.Style> styleSerializer) {
+        this.styleSerializer = styleSerializer;
+    }
+
+    @Override
+    public void write(Encoder encoder, StyledTextOutputEvent.Span value) throws Exception {
+        styleSerializer.write(encoder, value.getStyle());
+        encoder.writeString(value.getText());
+    }
+
+    @Override
+    public StyledTextOutputEvent.Span read(Decoder decoder) throws Exception {
+        return new StyledTextOutputEvent.Span(styleSerializer.read(decoder), decoder.readString());
+    }
+}
